agentHost: move CodexSessionConfigKey to common/ layer
Mirrors the layout used for Claude (`common/claudeSessionConfigKeys.ts`)
so that browser-side pickers can consume the Codex session-config key
strings + narrow helpers without violating the platform layering rule
that forbids `common/` importing from `node/`.

- New canonical file: `src/vs/platform/agentHost/common/codexSessionConfigKeys.ts`
  Exports `CodexSessionConfigKey`, `CodexApprovalPolicy`, `CodexSandboxMode`,
  `CodexWebSearchMode`, `CodexReasoningEffort` (string-literal unions
  structurally identical to the auto-generated protocol types), plus
  the `narrow*`, `isCodexSupportedModel`, and `normalizeCodexModelId`
  helpers. The protocol-generated types stay where they are; we keep
  parallel string unions in `common/` so the picker code never needs
  to reach into `node/`.
- The existing `node/codex/codexSessionConfigKeys.ts` is reduced to a
  re-export shim so all existing imports (`codexAgent.ts`, tests, etc.)
  keep working unchanged.

Validated with `compile-check-ts-native` and `valid-layers-check`.

Refs #cff62b58-fced-4fd2-9b33-f6f5eddd1bca
